Completed homework 3

+++ Lecture6/Homework2_sumAverage.java
@@ -8,15 +8,15 @@ public class Homework2_sumAverage {
         Scanner in = new Scanner(System.in);
 
         System.out.print("Въведете елемент 1: ");
-        double element1 = in.nextInt();
+        int element1 = in.nextInt();
 
         System.out.print("Въведете елемент 2: ");
-        double element2 = in.nextInt();
+        int element2 = in.nextInt();
 
         System.out.print("Въведете елемент 3: ");
-        double element3 = in.nextInt();
+        int element3 = in.nextInt();
 
-        double[] array = {element1, element2, element3};
+        int[] array = {element1, element2, element3};
         double sumAverage = array[0];
         double sum = 0;
         for (int i = array.length - 1; i >= 0; i--) {
